move message conversion into providers
